+++ packages/test-snaps/src/features/snaps/get-entropy/components/index.ts
@@ -1,1 +1,2 @@
+export * from './EntropySelector';
 export * from './SignMessage';